refactor : preview API를 CompletableFuture로 대체 (#110)
* refactor : preview API 대신 ComputableFuture 사용

* chore: Java 스타일 수정

* refactor : preview API 관련 설정 제거

* refactor : 각각의 대기 주문에 @Transactional이 적용되지 않는 문제 해결

* chore: Java 스타일 수정

---------
